+++ lib/conjur/command.rb
@@ -266,6 +266,22 @@ an alternative destination role.)
           obj.role.revoke_from member
         end
       end
+
+      def retire_internal_role roleObj
+        members = roleObj.members
+        # Move the invoking role to the end of the roles list, so that it doesn't
+        # lose its permissions in the middle of this operation.
+        self_member = members.select{|m| m.member.roleid == current_user.role.roleid}
+        self_member.each do |m|
+          members.delete m
+        end
+        members.concat self_member if self_member
+        members.each do |r|
+          member = api.role(r.member)
+          puts "Revoking from role #{member.roleid}"
+          roleObj.revoke_from member
+        end
+      end
       
       def give_away_resource obj, options
         destination = options[:"destination-role"]
@@ -432,6 +448,14 @@ an alternative destination role.)
         
         password
       end
+      
+      def has_admin?(role, other_role)
+        memberships = role.memberships.map(&:roleid)
+        other_role.members.any? { |m| memberships.member?(m.member.roleid) && m.admin_option }
+      rescue RestClient::Forbidden
+        false
+      end
+
     end
   end
 end